backend/requirements.txt
fastapi
uvicorn
boto3
requests
openai
google-genai
Pillow
